refactor: deployment e2e
